mqtt-auth/requirements.txt
psycopg2-binary==2.9.9
flask==3.1.0
